Added php protobuf extension. server.py improvements

+++ Makefile
@@ -2,25 +2,18 @@
 help:
 	@grep -E '(^[a-zA-Z_-]+:.*?##.*$$)|(^##)' $(MAKEFILE_LIST) | awk 'BEGIN {FS = ":.*?## "}{printf "\033[32m%-30s\033[0m %s\n", $$1, $$2}' | sed -e 's/\[32m##/[33m/'
 
-init:
+install: venv dist build ## Do everything
+
+venv: ## Install python virtual environment
 	python3 -m venv .venv
-	.venv/bin/pip3 install -r requirements.txt
+	.venv/bin/pip3 install --upgrade pip -r requirements.txt
 
-build: dir-vars-passed dist-copy ## Build server, "--mount" and "--volume" required
-	./.venv/bin/python server.py build --mount ${mount} --volume ${volume}
+dist: dist-exists ## Derive config files from .dist examples
+	test -f .env || cp config/dist/.env.dist .env
+	test -f config/docker-compose.template.yml || cp config/dist/docker-compose.template.yml.dist config/docker-compose.template.yml
+	test -f config/sites.yml || cp config/dist/sites.yml.dist config/sites.yml
 
-dist-copy: dist-exist ## Derive config files from .dist examples
-	cp config/dist/.env.dist config/.env
-	cp config/dist/docker-compose.template.yml.dist config/docker-compose.template.yml
-	cp config/dist/sites.yml.dist config/sites.yml
+build: ## Build server
+	./.venv/bin/python server.py build
 
-
-dist-exist: config/dist/.env.dist config/dist/docker-compose.template.yml.dist config/dist/sites.yml.dist
-
-dir-vars-passed:
-ifndef mount
-	$(error mount is undefined)
-endif
-ifndef volume
-	$(error volume is undefined)
-endif
+dist-exists: config/dist/.env.dist config/dist/docker-compose.template.yml.dist config/dist/sites.yml.dist
